agents/coordinator/requirements.txt
strands-agents
bedrock-agentcore
boto3>=1.34.0
